improve UI of arranging products in search page
