--- src/view/search_product/RoundedPanel.java
@@ -1,30 +1,0 @@
-package view.search_product;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class RoundedPanel extends JPanel {
-    private int cornerRadius;
-
-    public RoundedPanel(int radius) {
-        super();
-        cornerRadius = radius;
-        setOpaque(false);
-    }
-
-    @Override
-    protected void paintComponent(Graphics g) {
-        super.paintComponent(g);
-        Dimension arcs = new Dimension(cornerRadius, cornerRadius);
-        int width = getWidth();
-        int height = getHeight();
-        Graphics2D graphics = (Graphics2D) g;
-        graphics.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-
-        // Draws the rounded panel with borders.
-        graphics.setColor(getBackground());
-        graphics.fillRoundRect(0, 0, width - 1, height - 1, arcs.width, arcs.height);
-        graphics.setColor(getForeground());
-        graphics.drawRoundRect(0, 0, width - 1, height - 1, arcs.width, arcs.height);
-    }
-}
